api/requirements.txt
fastapi==0.115.6
Pillow==11.0.0
uvicorn==0.34.0
tensorflow==2.18.0
python-multipart==0.0.19